Handle files that are already copied but not deleted

+++ lib/Handler/HandlerFilesystem.php
@@ -36,14 +36,21 @@ class HandlerFilesystem extends HandlerBase
         if ($this->descriptor->keep) {
             copy($this->descriptor->source,$destination);
         } else {
-            try {
-                if (!rename($this->descriptor->source,$destination)) {
+            if (file_exists($destination)) {
+                $this->error("File '".$this->descriptor->source."' already in originals.");
+            } else {
+                try {
+                    if (!rename($this->descriptor->source,$destination)) {
+                        copy($this->descriptor->source,$destination);
+                        unlink($this->descriptor->source);   
+                    }
+                } catch (\Exception $e) {
                     copy($this->descriptor->source,$destination);
-                    unlink($this->descriptor->source);   
+                    unlink($this->descriptor->source);
+                    if (file_exists($this->descriptor->source)) {
+                        $this->error("File '".$this->descriptor->source."' could not be deleted. Is kept.")
+                    }
                 }
-            } catch (\Exception $e) {
-                copy($this->descriptor->source,$destination);
-                unlink($this->descriptor->source);               
             }
         }
         $this->descriptor->destination = $destination;
